Tryfix windows test. Includes module.
